OrderApp
The OrderApp console application allows the user to create and manage orders with different products, such as a laptop, mouse or keyboard. The user can add and remove products from the order, as well as view the current value of the order with automatically calculated discounts. The available discounts include: 10% discount on the second cheapest product, 20% on the third cheapest, 5% on orders over PLN 5,000 and 15% on the "Laptop + Mouse" set. The application presents details of the discounts applied, which allows you to track savings in real time.